requirements.txt
azure-identity>=1.16.0,<2.0.0
azure-mgmt-resource>=23.0.0,<24.0.0
python-dotenv>=1.0.0,<2.0.0
azure-mgmt-compute>=33.0.0,<34.0.0

